backend/requirements.txt
# requirements.txt

# FastAPI and Web Server
fastapi==0.115.8
uvicorn==0.34.0
starlette==0.45.3

# HTTP and Networking
h11==0.14.0
anyio==4.8.0
sniffio==1.3.1
httpx==0.27.0  # Required for FastAPI TestClient
python-multipart==0.0.9  # Required for form data handling

# Data Validation and Type Checking
pydantic==2.10.6
pydantic-settings==2.2.1
pydantic_core==2.27.2
typing_extensions==4.12.2
annotated-types==0.7.0
email-validator==2.1.1

# Configuration and Environment
python-dotenv==1.0.1
configparser==7.1.0

# Utilities
click==8.1.8
colorama==0.4.6
idna==3.10
PyJWT==2.8.0
bcrypt==4.1.2
passlib==1.7.4

# Database
pymongo==4.5.0
motor==3.3.1
sqlalchemy==2.0.28

# Web3 and Blockchain
web3==6.16.0
eth-typing==4.0.0
eth-utils==2.3.1
eth-account==0.11.0

# Cryptography
cryptography==42.0.5
py-ecc==7.0.0
pycryptodome==3.20.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
